Move types associated with workflow
